Wrap both JS files in combined Node/IIFE modules
The source of `math-perf.js` is now fully portable between the browser
and Node. It exports three properties either under the `MathPerf` global
object or as a CommonJS module:
 - recentOutputs: recently output number arrays for debugging
 - tests: an object the source for each test (moved from view.js)
 - runTest: takes a string for the test name and input and runs test loops

As a consequence of this commit, `view.js` will no longer expose any
global objects.

Also times are now in nanoseconds not microseconds.

Signed-off-by: Zac Delventhal <delventhalz@gmail.com>